refactor: template/Main to template/MiddleSection & MainHeader to MiddleSectionHeader
