+++ front/src/templates/MiddleSection/index.tsx
